requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
torch==2.0.1+cpu
torchvision==0.15.2+cpu
ultralytics==8.3.170
pillow==11.1.0
opencv-python-headless==4.11.0.86
jinja2==3.1.5
-f https://download.pytorch.org/whl/cpu/torch_stable.html
python-multipart
numpy==1.24.4

